feat(protocol): enable ETH deposit from L1 to L2 using withdrawalsRoot [1] (#13617)

+++ packages/protocol/contracts/L1/libs/LibProving.sol
@@ -57,7 +57,7 @@ library LibProving {
         bytes32 parentHash = oracles.parentHash;
         uint32 parentGasUsed = oracles.parentGasUsed;
 
-        for (uint i = 0; i < oracles.blks.length; ) {
+        for (uint256 i = 0; i < oracles.blks.length; ) {
             uint256 id = blockId + i;
 
             if (id <= state.lastVerifiedBlockId || id >= state.numBlocks)
